refactor: move the playback clock into a hook (#66)
Second of the three. The rAF loop drives everything that moves - canvas, playhead, audio element,
video overlay, frame prefetcher - at three deliberately different rates, and export runs through
the same loop, which is why it comes along rather than staying beside the recorder.

The hook owns six things App used to: isPlaying and currentTime, and the four refs the loop reads
instead of state so it never runs on a stale closure. Everything else is an input. There are a
lot of those, and they are grouped rather than listed flat, because a playback loop genuinely is
a function of the timeline, the media and where to paint - seventeen arguments in a row would be
hiding that rather than admitting it.

paintFrameRef moved up to sit with the other refs. It was declared six hundred lines below the
loop that reads it, which is what forced the hook call below everything; a ref has no
dependencies, so where it is declared is a readability choice and that one was costing something.

On the hook baseline: it grew by two and both are accounted for. One is the loop's existing
warning, now attributed to usePlayback.js instead of App.jsx. The other is new in App and is a
linter-visibility artefact rather than new risk: exhaustive-deps knows a useState setter is
stable and excludes it, but a value destructured from an unknown custom hook is not known-stable,
so `restore` becomes a reported dependency of the crash-recovery effect. Confirmed by probe -
swapping in a setter the linter can prove stable takes the count straight back to ten. The effect
runs once on mount, where the first-render closure is the correct one, and adding `restore` to
its dependency list would re-run crash recovery on every render.

App.jsx 4358 -> 4251. 371 tests.
